[1.4][sfSympalPlugin][1.0] Fixing php error
git-svn-id: http://svn.symfony-project.org/plugins/sfSympalPlugin/trunk@26914 ee427ae8-e902-0410-961c-c3ed070cd9f9

+++ lib/cache/sfSympalCache.class.php
@@ -53,7 +53,11 @@ class sfSympalCache
   public function resetRouteCache()
   {
     $cachePath = sfConfig::get('sf_cache_dir').'/'.sfConfig::get('sf_app').'/'.sfConfig::get('sf_environment').'/routes.cache.yml';
-    unlink($cachePath);
+    if (file_exists($cachePath))
+    {
+      unlink($cachePath);
+    }
+
     $context = sfContext::getInstance();
     $configCache = $context->getConfigCache();
